# server/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: MesaLista Wedding Registry API
  description: API for managing wedding gift registries, users, gifts, and purchases
  version: 1.0.0
  contact:
    name: MesaLista Support
    email: support@mesalista.com

servers:
  - url: http://localhost:5001/api
    description: Development server
  - url: https://mesalista-production.up.railway.app/api
    description: Production server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    # User Schemas
    User:
      type: object
      required:
        - firstName
        - lastName
        - email
        - password
      properties:
        id:
          type: integer
          description: The auto-generated id of the user
        firstName:
          type: string
          description: First name of the user
        lastName:
          type: string
          nullable: true
          description: Last name of the user
        email:
          type: string
          format: email
          description: The email address of the user
        password:
          type: string
          format: password
          description: The password of the user (hashed in responses)
        phoneNumber:
          type: string
          nullable: true
          description: Phone number of the user
        role:
          type: string
          enum: [USER, ADMIN, COUPLE]
          description: The role of the user
        imageUrl:
          type: string
          nullable: true
          description: Profile image URL
        spouseFirstName:
          type: string
          nullable: true
          description: Spouse's first name
        spouseLastName:
          type: string
          nullable: true
          description: Spouse's last name
        createdAt:
          type: string
          format: date-time
          description: The date the user was created
        updatedAt:
          type: string
          format: date-time
          description: The date the user was last updated

    UserResponse:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        email:
          type: string
        phoneNumber:
          type: string
          nullable: true
        role:
          type: string
        imageUrl:
          type: string
          nullable: true
        spouseFirstName:
          type: string
          nullable: true
        spouseLastName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password

    LoginResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserResponse'
        token:
          type: string
          description: JWT authentication token

    # Gift Schemas
    Gift:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        category:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        quantity:
          type: integer
        isPurchased:
          type: boolean
        isMostWanted:
          type: boolean
        weddingListId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateGiftRequest:
      type: object
      required:
        - title
        - price
        - weddingListId
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: string
        quantity:
          type: integer
          default: 1
        isMostWanted:
          type: boolean
          default: false
        weddingListId:
          type: integer

    # Wedding List Schemas
    WeddingList:
      type: object
      properties:
        id:
          type: integer
        coupleName:
          type: string
        weddingDate:
          type: string
          format: date-time
        weddingLocation:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    WeddingListWithGifts:
      allOf:
        - $ref: '#/components/schemas/WeddingList'
        - type: object
          properties:
            gifts:
              type: array
              items:
                $ref: '#/components/schemas/Gift'

    # Cart Schemas
    Cart:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'

    CartItem:
      type: object
      properties:
        id:
          type: integer
        cartId:
          type: integer
        giftId:
          type: integer
        quantity:
          type: integer
        gift:
          $ref: '#/components/schemas/Gift'

    AddToCartRequest:
      type: object
      required:
        - giftId
        - quantity
      properties:
        giftId:
          type: integer
        quantity:
          type: integer
          minimum: 1

    # Payment Schemas
    Payment:
      type: object
      properties:
        id:
          type: integer
        amount:
          type: number
        currency:
          type: string
          default: "usd"
        status:
          type: string
          enum: [PENDING, COMPLETED, FAILED, CANCELLED]
        stripePaymentIntentId:
          type: string
        userId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    PaymentRequest:
      type: object
      required:
        - amount
        - currency
      properties:
        amount:
          type: number
          description: Amount in cents
        currency:
          type: string
          default: "usd"

    # Purchase Schemas
    Purchase:
      type: object
      properties:
        id:
          type: integer
        giftId:
          type: integer
        userId:
          type: integer
        quantity:
          type: integer
        totalAmount:
          type: number
        status:
          type: string
          enum: [PENDING, DELIVERED, THANKED]
        message:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string

paths:
  # User endpoints
  /users:
    post:
      summary: Register a new user
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - firstName
                - email
                - password
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                phoneNumber:
                  type: string
                role:
                  type: string
                  enum: [USER, ADMIN, COUPLE]
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Missing required fields or invalid data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: Get all users
      tags: [Users]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of all users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /users/login:
    post:
      summary: Login a user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /users/me:
    get:
      summary: Get current authenticated user
      tags: [Users]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /users/{id}:
    get:
      summary: Get user by ID
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the user
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update a user
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the user to update
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                phoneNumber:
                  type: string
                role:
                  type: string
                  enum: [USER, ADMIN, COUPLE]
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete a user
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the user to delete
      responses:
        '200':
          description: User deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Gift endpoints
  /gifts:
    get:
      summary: Get all gifts
      tags: [Gifts]
      responses:
        '200':
          description: List of all gifts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Gift'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a new gift
      tags: [Gifts]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGiftRequest'
      responses:
        '201':
          description: Gift created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gift'
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /gifts/{id}:
    get:
      summary: Get gift by ID
      tags: [Gifts]
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the gift
      responses:
        '200':
          description: Gift details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gift'
        '404':
          description: Gift not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update a gift
      tags: [Gifts]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the gift to update
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                price:
                  type: number
                category:
                  type: string
                imageUrl:
                  type: string
                quantity:
                  type: integer
                isMostWanted:
                  type: boolean
      responses:
        '200':
          description: Gift updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gift'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Gift not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete a gift
      tags: [Gifts]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the gift to delete
      responses:
        '200':
          description: Gift deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Gift not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Cart endpoints
  /cart:
    get:
      summary: Get user's cart
      tags: [Cart]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User's cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Add item to cart
      tags: [Cart]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddToCartRequest'
      responses:
        '200':
          description: Item added to cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Clear cart
      tags: [Cart]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Cart cleared successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /cart/{id}/checkout:
    post:
      summary: Checkout cart
      tags: [Cart]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Cart ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - inviteeName
                - message
              properties:
                inviteeName:
                  type: string
                message:
                  type: string
      responses:
        '200':
          description: Checkout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  purchaseIds:
                    type: array
                    items:
                      type: integer
                  message:
                    type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Cart not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Payment endpoints
  /payments:
    get:
      summary: Get all payments
      tags: [Payments]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Payment'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a payment
      tags: [Payments]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '201':
          description: Payment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /payments/{id}:
    get:
      summary: Get payment by ID
      tags: [Payments]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Payment ID
      responses:
        '200':
          description: Payment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

tags:
  - name: Users
    description: User management operations
  - name: Authentication
    description: Authentication operations
  - name: Gifts
    description: Gift management operations
  - name: Cart
    description: Shopping cart operations
  - name: Payments
    description: Payment processing operations